Move EffectUtility to a file not referenced by WindowsGL
